requirements.txt
discord.py
aiomysql
requests
emoji
Pillow
python-dotenv
